Selection System, Enemy Death, Attack and stunning systems implemented
Players can now select which enemies to attack. Once the attack button is pressed, the script refers to a combat selection component to filter the combatants list to just enemies. The player can then use the A and D keys to move between available enemies. The player can then press the selection button to select and damage an enemy. Enemies can now die and attack each turn. When enemies die,
 the next queue turn state will remove them from the list and a Character Death script attached to the enemy will disable their mesh. When an enemy enters the turn state, they immediately call an attack method from an Enemy Attack script at random. This list is filtered from available players. Characters can now be stunned, delaying their turn. After the player runs out of stun health they are moved back three spaces in the queue turn. And their stun health is reset.